Move the gridfm read round trips to the facade tests
The facade owns the Parquet reader, and the matrix crate cannot see the
facade without making the crate graph circular, so the fourteen write
to read round trip tests move into powerio/tests/gridfm_read.rs over
the existing dev dependency. The matrix module doc names the facade
readers as plain paths, and the facade reader doc names the matrix
writer the same way, since neither crate can doc link across that
boundary.
